src/requirements.txt
boto3
pyyaml
termcolor
ctransformers
ctransformers[gptq]
tqdm